🩹 chore: 페이지 파일에서 정의한 상수 정리
