+++ app/(route)/signup/_components/SearchArtist.tsx
@@ -1,4 +1,4 @@
-import { ArtistType, MOCK } from "@/(route)/setting/favorite/page";
+import { ArtistType, MOCK } from "app/_constants/mock";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import ArtistCard from "@/components/ArtistCard";
