+++ app/_constants/mock.ts
@@ -1,0 +1,111 @@
+export type ArtistType = {
+  name: string;
+  group?: string[];
+  profileImage: string;
+};
+
+export const MOCK: ArtistType[] = [
+  {
+    name: "마크",
+    group: ["NCT", "NCT127", "NCT DREAM"],
+    profileImage: "",
+  },
+  {
+    name: "도영",
+    group: ["NCT", "NCT127"],
+    profileImage: "",
+  },
+  {
+    name: "켄지",
+    profileImage: "",
+  },
+  {
+    name: "소희",
+    group: ["원더걸스"],
+    profileImage: "",
+  },
+  {
+    name: "나비스",
+    profileImage: "",
+  },
+  {
+    name: "윈터",
+    group: ["aespa"],
+    profileImage: "",
+  },
+  {
+    name: "수호",
+    group: ["exo"],
+    profileImage: "",
+  },
+  {
+    name: "디오",
+    group: ["exo"],
+    profileImage: "",
+  },
+  {
+    name: "수영",
+    group: ["소녀시대"],
+    profileImage: "",
+  },
+  {
+    name: "태연",
+    group: ["소녀시대"],
+    profileImage: "",
+  },
+  {
+    name: "동해",
+    group: ["슈퍼주니어"],
+    profileImage: "",
+  },
+  {
+    name: "키",
+    group: ["샤이니"],
+    profileImage: "",
+  },
+  {
+    name: "민호",
+    group: ["샤이니"],
+    profileImage: "",
+  },
+  {
+    name: "조이",
+    group: ["레드벨벳"],
+    profileImage: "",
+  },
+  {
+    name: "아이린",
+    group: ["레드벨벳"],
+    profileImage: "",
+  },
+  {
+    name: "효연",
+    group: ["소녀시대"],
+    profileImage: "",
+  },
+  {
+    name: "웬디",
+    group: ["레드벨벳"],
+    profileImage: "",
+  },
+  {
+    name: "카리나",
+    group: ["aespa"],
+    profileImage: "",
+  },
+  {
+    name: "지젤",
+    group: ["aespa"],
+    profileImage: "",
+  },
+  {
+    name: "유리",
+    group: ["소녀시대"],
+    profileImage: "",
+  },
+  {
+    name: "써니",
+    group: ["소녀시대"],
+    profileImage: "",
+  },
+];
